@@ -14,11 +14,11 @@ dependencies:
   intl: ^0.19.0
   flutter_riverpod: ^2.6.1
   go_router: ^15.1.1
-  flutter_svg: ^2.0.17
+  flutter_svg: ^2.1.0
   cached_network_image: ^3.4.1
   flutter_secure_storage: ^9.2.4
   shared_preferences: ^2.5.3
-  http: ^1.3.0
+  http: ^1.4.0
   json_annotation: ^4.9.0
   freezed_annotation: ^3.0.0
   geolocator: ^14.0.0
@@ -27,17 +27,17 @@ dependencies:
   flutter_launcher_icons: ^0.14.3
   flutter_dotenv: ^5.2.1
   logger: ^2.0.2+1
-  connectivity_plus: ^6.1.4
+  connectivity_plus: ^6.0.4
   flutter_animate: ^4.5.2
   flutter_hooks: ^0.21.2
   hooks_riverpod: ^2.6.1
-  mapbox_maps_flutter: ^2.7.0
-  file_selector: ^0.9.3+1
+  mapbox_maps_flutter: ^2.8.0
+  file_selector: ^1.0.3
   flutter_localizations:
     sdk: flutter
   file_picker: ^10.1.2
-  device_info_plus: ^9.1.2
-  battery_plus: ^5.0.3
+  device_info_plus: ^11.4.0
+  battery_plus: ^6.2.1
   url_launcher: ^6.3.1
   package_info_plus: ^8.3.0
 
