@@ -3,6 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../widgets/shared_filter_widgets.dart';
 import '../../../../core/widgets/segmented_tab_control.dart';
 import 'package:travloc/core/widgets/preference_tile.dart';
+import '../widgets/guide_card.dart';
 
 final guideFiltersProvider =
     StateNotifierProvider<GuideFilters, Map<String, dynamic>>((ref) {
@@ -50,7 +51,7 @@ class TravelBuddiesFilters extends StateNotifier<Map<String, dynamic>> {
   TravelBuddiesFilters()
     : super({
         'searchQuery': '',
-        'travelDates': null, // Placeholder for date range or similar
+        'travelDates': null,
         'interests': <String>{},
         'languages': <String>{},
       });
@@ -68,12 +69,12 @@ class GuidesScreen extends ConsumerStatefulWidget {
 }
 
 class _GuidesScreenState extends ConsumerState<GuidesScreen>
-    with SingleTickerProviderStateMixin {
-  late TabController _tabController;
-  final List<String> _tabs = ['Guides', 'Travel Buddies'];
+    with SingleTickerProviderStateMixin, AutomaticKeepAliveClientMixin {
+  late final TabController _tabController;
+  final List<String> _tabs = const ['Guides', 'Travel Buddies'];
   final TextEditingController _guideSearchController = TextEditingController();
   final TextEditingController _buddySearchController = TextEditingController();
-  final List<String> _availableLanguages = [
+  final List<String> _availableLanguages = const [
     'English',
     'Spanish',
     'French',
@@ -82,6 +83,9 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
     'Chinese',
     'Japanese',
   ];
+
+  @override
+  bool get wantKeepAlive => true;
 
   @override
   void initState() {
@@ -104,56 +108,99 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
   }
 
   void _showGuideFilterDialog(BuildContext context) {
+    final filters = ref.read(guideFiltersProvider);
     showDialog(
       context: context,
       builder:
-          (context) => SharedFilterDialog(
-            title: 'Filter Guides',
-            filterOptions: [
-              PreferenceTile(
-                icon: Icons.monetization_on,
-                title: 'Price Range',
-                onTap: () {
-                  Navigator.pop(context);
-                  _showPriceRangePicker(context);
-                },
-                trailing: const Icon(
-                  Icons.arrow_forward_ios,
-                  color: Colors.black,
-                ),
-                backgroundColor: const Color(0xFFF3EDFF),
-                iconBackgroundColor: const Color(0xFFD1C4E9),
+          (context) => AlertDialog(
+            title: const Text('Filter Guides'),
+            content: SingleChildScrollView(
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  PreferenceTile(
+                    icon: Icons.monetization_on,
+                    title: 'Price Range',
+                    subtitle:
+                        '\$${filters['priceRange'].start.round()} - \$${filters['priceRange'].end.round()}',
+                    onTap: () => _showPriceRangePicker(context),
+                  ),
+                  PreferenceTile(
+                    icon: Icons.star,
+                    title: 'Minimum Rating',
+                    subtitle: '${filters['minRating']} stars',
+                    onTap: () => _showRatingFilter(context),
+                  ),
+                  PreferenceTile(
+                    icon: Icons.language,
+                    title: 'Languages',
+                    subtitle:
+                        (filters['languages'] as Set<String>).isEmpty
+                            ? 'Any'
+                            : (filters['languages'] as Set<String>).join(', '),
+                    onTap: () => _showLanguageFilter(context),
+                  ),
+                ],
               ),
-              PreferenceTile(
-                icon: Icons.star,
-                title: 'Rating',
-                onTap: () {
-                  Navigator.pop(context);
-                  _showRatingFilter(context);
-                },
-                trailing: const Icon(
-                  Icons.arrow_forward_ios,
-                  color: Colors.black,
-                ),
-                backgroundColor: const Color(0xFFF3EDFF),
-                iconBackgroundColor: const Color(0xFFD1C4E9),
-              ),
-              PreferenceTile(
-                icon: Icons.language,
-                title: 'Languages',
-                onTap: () {
-                  Navigator.pop(context);
-                  _showLanguageFilter(context);
-                },
-                trailing: const Icon(
-                  Icons.arrow_forward_ios,
-                  color: Colors.black,
-                ),
-                backgroundColor: const Color(0xFFF3EDFF),
-                iconBackgroundColor: const Color(0xFFD1C4E9),
+            ),
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.pop(context),
+                child: const Text('Done'),
               ),
             ],
-            onApply: () => Navigator.pop(context),
+          ),
+    );
+  }
+
+  void _showBuddyFilterDialog(BuildContext context) {
+    final filters = ref.read(travelBuddiesFiltersProvider);
+    showDialog(
+      context: context,
+      builder:
+          (context) => AlertDialog(
+            title: const Text('Filter Travel Buddies'),
+            content: SingleChildScrollView(
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  PreferenceTile(
+                    icon: Icons.calendar_today,
+                    title: 'Travel Dates',
+                    subtitle: filters['travelDates'] ?? 'Any',
+                    onTap: () {
+                      // Implement date picker
+                    },
+                  ),
+                  PreferenceTile(
+                    icon: Icons.interests,
+                    title: 'Interests',
+                    subtitle:
+                        (filters['interests'] as Set<String>).isEmpty
+                            ? 'Any'
+                            : (filters['interests'] as Set<String>).join(', '),
+                    onTap: () {
+                      // Implement interests picker
+                    },
+                  ),
+                  PreferenceTile(
+                    icon: Icons.language,
+                    title: 'Languages',
+                    subtitle:
+                        (filters['languages'] as Set<String>).isEmpty
+                            ? 'Any'
+                            : (filters['languages'] as Set<String>).join(', '),
+                    onTap: () => _showLanguageFilter(context),
+                  ),
+                ],
+              ),
+            ),
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.pop(context),
+                child: const Text('Done'),
+              ),
+            ],
           ),
     );
   }
@@ -171,8 +218,8 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
               max: 1000,
               divisions: 20,
               labels: RangeLabels(
-                ' 24${filters['priceRange'].start.round()}',
-                ' 24${filters['priceRange'].end.round()}',
+                '\$${filters['priceRange'].start.round()}',
+                '\$${filters['priceRange'].end.round()}',
               ),
               onChanged: (RangeValues values) {
                 ref
@@ -210,7 +257,10 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
                     ref.read(guideFiltersProvider.notifier).updateRating(value);
                   },
                 ),
-                Text('${filters['minRating']} stars'),
+                Text(
+                  '${filters['minRating']} stars',
+                  style: const TextStyle(fontSize: 16),
+                ),
               ],
             ),
             actions: [
@@ -259,344 +309,68 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
     );
   }
 
-  void _showBuddyFilterDialog(BuildContext context) {
-    showDialog(
-      context: context,
-      builder:
-          (context) => SharedFilterDialog(
-            title: 'Filter Buddies',
-            filterOptions: [
-              PreferenceTile(
-                icon: Icons.date_range,
-                title: 'Travel Dates',
-                onTap: () {
-                  // Implement date filter
+  Widget _buildTabContent(int tabIndex) {
+    return CustomScrollView(
+      physics: const BouncingScrollPhysics(),
+      keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
+      slivers: [
+        SliverPersistentHeader(
+          floating: true,
+          delegate: _SliverSearchBarDelegate(
+            child: Padding(
+              padding: const EdgeInsets.all(12),
+              child: SharedSearchBar(
+                controller:
+                    tabIndex == 0
+                        ? _guideSearchController
+                        : _buddySearchController,
+                hintText:
+                    tabIndex == 0
+                        ? 'Search guides...'
+                        : 'Search travel buddies...',
+                onChanged: (value) {
+                  if (tabIndex == 0) {
+                    ref
+                        .read(guideFiltersProvider.notifier)
+                        .updateSearchQuery(value);
+                  } else {
+                    ref
+                        .read(travelBuddiesFiltersProvider.notifier)
+                        .updateSearchQuery(value);
+                  }
                 },
-                trailing: const Icon(
-                  Icons.arrow_forward_ios,
-                  color: Colors.black,
-                ),
-                backgroundColor: const Color(0xFFF3EDFF),
-                iconBackgroundColor: const Color(0xFFD1C4E9),
+                backgroundColor: Colors.white,
+                iconColor: Colors.black54,
+                borderRadius: 16,
               ),
-              PreferenceTile(
-                icon: Icons.interests,
-                title: 'Interests',
-                onTap: () {
-                  // Implement interests filter
-                },
-                trailing: const Icon(
-                  Icons.arrow_forward_ios,
-                  color: Colors.black,
-                ),
-                backgroundColor: const Color(0xFFF3EDFF),
-                iconBackgroundColor: const Color(0xFFD1C4E9),
-              ),
-              PreferenceTile(
-                icon: Icons.language,
-                title: 'Languages',
-                onTap: () {
-                  // Implement languages filter
-                },
-                trailing: const Icon(
-                  Icons.arrow_forward_ios,
-                  color: Colors.black,
-                ),
-                backgroundColor: const Color(0xFFF3EDFF),
-                iconBackgroundColor: const Color(0xFFD1C4E9),
-              ),
-            ],
-            onApply: () => Navigator.pop(context),
-          ),
-    );
-  }
-
-  Widget _buildGuideCard(BuildContext context, int index) {
-    return Container(
-      margin: const EdgeInsets.fromLTRB(6, 0, 6, 6),
-      decoration: BoxDecoration(
-        color: const Color(0xFFF3EDFF), // pastel purple
-        borderRadius: BorderRadius.circular(20),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withAlpha((0.08 * 255).toInt()),
-            blurRadius: 8,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: InkWell(
-        borderRadius: BorderRadius.circular(20),
-        onTap: () {}, // Navigation logic can be added here
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              // Avatar
-              CircleAvatar(
-                radius: 24,
-                backgroundColor: const Color(0xFFD1C4E9),
-                child: const Text('🧑‍🎤', style: TextStyle(fontSize: 24)),
-              ),
-              const SizedBox(width: 14),
-              // Main Info
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: [
-                        Expanded(
-                          child: Text(
-                            'Guide Name',
-                            style: const TextStyle(
-                              fontWeight: FontWeight.bold,
-                              fontSize: 16,
-                              color: Colors.black,
-                              letterSpacing: 0.2,
-                            ),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ),
-                        const SizedBox(width: 8),
-                        IconButton(
-                          icon: Icon(
-                            Icons.favorite_border,
-                            color: Colors.black26,
-                            size: 22,
-                          ),
-                          onPressed: () {},
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 2),
-                    Row(
-                      children: [
-                        const Icon(Icons.star, size: 15, color: Colors.amber),
-                        const SizedBox(width: 3),
-                        Text(
-                          '4.8',
-                          style: TextStyle(
-                            color: Colors.grey[700],
-                            fontSize: 14,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                        const SizedBox(width: 3),
-                        Text(
-                          '(120)',
-                          style: TextStyle(
-                            color: Colors.grey[400],
-                            fontSize: 12,
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 6),
-                    Text(
-                      'Specializes in adventure tours and cultural experiences.',
-                      style: TextStyle(
-                        color: Colors.grey[800],
-                        fontSize: 13,
-                        fontWeight: FontWeight.w400,
-                      ),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    const SizedBox(height: 6),
-                    Row(
-                      children: [
-                        const Icon(
-                          Icons.location_on,
-                          size: 13,
-                          color: Colors.grey,
-                        ),
-                        const SizedBox(width: 3),
-                        Text(
-                          'Barcelona, Spain',
-                          style: TextStyle(
-                            color: Colors.grey[500],
-                            fontSize: 12,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-              // Book button
-              Padding(
-                padding: const EdgeInsets.only(left: 10, top: 2),
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.black,
-                    foregroundColor: Colors.white,
-                    minimumSize: const Size(56, 32),
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 12,
-                      vertical: 0,
-                    ),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    elevation: 0,
-                  ),
-                  onPressed: () {}, // Navigation logic can be added here
-                  child: const Text('Book', style: TextStyle(fontSize: 14)),
-                ),
-              ),
-            ],
+            ),
           ),
         ),
-      ),
-    );
-  }
-
-  Widget _buildBuddyCard(BuildContext context, int index) {
-    return Container(
-      margin: const EdgeInsets.fromLTRB(6, 0, 6, 6),
-      decoration: BoxDecoration(
-        color: const Color(0xFFF3EDFF), // pastel purple
-        borderRadius: BorderRadius.circular(20),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withAlpha((0.08 * 255).toInt()),
-            blurRadius: 8,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: InkWell(
-        borderRadius: BorderRadius.circular(20),
-        onTap: () {}, // Navigation logic can be added here
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              // Avatar
-              CircleAvatar(
-                radius: 24,
-                backgroundColor: const Color(0xFFD1C4E9),
-                child: const Text('🧑‍🤝‍🧑', style: TextStyle(fontSize: 24)),
+        SliverPadding(
+          padding: const EdgeInsets.only(bottom: 80),
+          sliver: SliverList(
+            delegate: SliverChildBuilderDelegate(
+              (context, index) => GuideCard(
+                key: ValueKey('guide_${tabIndex}_$index'),
+                index: index,
+                onTap: () {
+                  // Handle tap
+                },
+                onFavorite: () {
+                  // Handle favorite
+                },
               ),
-              const SizedBox(width: 14),
-              // Main Info
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: [
-                        Expanded(
-                          child: Text(
-                            'Travel Buddy Name',
-                            style: const TextStyle(
-                              fontWeight: FontWeight.bold,
-                              fontSize: 16,
-                              color: Colors.black,
-                              letterSpacing: 0.2,
-                            ),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ),
-                        const SizedBox(width: 8),
-                        IconButton(
-                          icon: Icon(
-                            Icons.favorite_border,
-                            color: Colors.black26,
-                            size: 22,
-                          ),
-                          onPressed: () {},
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 2),
-                    Row(
-                      children: [
-                        const Icon(
-                          Icons.location_on,
-                          size: 15,
-                          color: Colors.grey,
-                        ),
-                        const SizedBox(width: 3),
-                        Text(
-                          'Barcelona, Spain',
-                          style: TextStyle(
-                            color: Colors.grey[700],
-                            fontSize: 13,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 6),
-                    Text(
-                      'Interested in adventure tours and cultural experiences.',
-                      style: TextStyle(
-                        color: Colors.grey[800],
-                        fontSize: 13,
-                        fontWeight: FontWeight.w400,
-                      ),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    const SizedBox(height: 6),
-                    Row(
-                      children: [
-                        const Icon(
-                          Icons.calendar_today,
-                          size: 13,
-                          color: Colors.grey,
-                        ),
-                        const SizedBox(width: 3),
-                        Text(
-                          'Available: June 2024',
-                          style: TextStyle(
-                            color: Colors.grey[500],
-                            fontSize: 12,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-              // Connect button
-              Padding(
-                padding: const EdgeInsets.only(left: 10, top: 2),
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.black,
-                    foregroundColor: Colors.white,
-                    minimumSize: const Size(70, 32),
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 12,
-                      vertical: 0,
-                    ),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    elevation: 0,
-                  ),
-                  onPressed: () {}, // Navigation logic can be added here
-                  child: const Text('Connect', style: TextStyle(fontSize: 14)),
-                ),
-              ),
-            ],
+              childCount: 10,
+            ),
           ),
         ),
-      ),
+      ],
     );
   }
 
   @override
   Widget build(BuildContext context) {
+    super.build(context);
     return Stack(
       children: [
         Container(
@@ -606,61 +380,10 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
               Expanded(
                 child: TabBarView(
                   controller: _tabController,
-                  children: [
-                    // Guides Tab
-                    ListView.builder(
-                      padding: EdgeInsets.zero,
-                      itemCount: 11, // 1 for search bar + 10 cards
-                      itemBuilder: (context, index) {
-                        if (index == 0) {
-                          return Padding(
-                            padding: const EdgeInsets.all(12),
-                            child: SharedSearchBar(
-                              controller: _guideSearchController,
-                              hintText: 'Search guides...',
-                              onChanged: (value) {
-                                ref
-                                    .read(guideFiltersProvider.notifier)
-                                    .updateSearchQuery(value);
-                              },
-                              backgroundColor: Colors.white,
-                              iconColor: Colors.black54,
-                              borderRadius: 16,
-                            ),
-                          );
-                        }
-                        return _buildGuideCard(context, index - 1);
-                      },
-                    ),
-                    // Travel Buddies Tab
-                    ListView.builder(
-                      padding: EdgeInsets.zero,
-                      itemCount: 11, // 1 for search bar + 10 cards
-                      itemBuilder: (context, index) {
-                        if (index == 0) {
-                          return Padding(
-                            padding: const EdgeInsets.all(12),
-                            child: SharedSearchBar(
-                              controller: _buddySearchController,
-                              hintText: 'Search travel buddies...',
-                              onChanged: (value) {
-                                ref
-                                    .read(travelBuddiesFiltersProvider.notifier)
-                                    .updateSearchQuery(value);
-                              },
-                              backgroundColor: Colors.white,
-                              iconColor: Colors.black54,
-                              borderRadius: 16,
-                            ),
-                          );
-                        }
-                        return _buildBuddyCard(context, index - 1);
-                      },
-                    ),
-                  ],
+                  physics: const BouncingScrollPhysics(),
+                  children: [_buildTabContent(0), _buildTabContent(1)],
                 ),
               ),
-              // Animated SegmentedTabControl at Bottom
               SafeArea(
                 top: false,
                 minimum: const EdgeInsets.only(bottom: 8),
@@ -674,7 +397,7 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
                         _tabController.index.toDouble(),
                     onTabSelected: (index) {
                       setState(() {
-                        _tabController.index = index;
+                        _tabController.animateTo(index);
                       });
                     },
                   ),
@@ -683,7 +406,6 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
             ],
           ),
         ),
-        // Floating Filter Button (overlays both tabs)
         Positioned(
           bottom: 70,
           right: 24,
@@ -702,5 +424,32 @@ class _GuidesScreenState extends ConsumerState<GuidesScreen>
         ),
       ],
     );
+  }
+}
+
+class _SliverSearchBarDelegate extends SliverPersistentHeaderDelegate {
+  final Widget child;
+  static const double height = 68;
+
+  _SliverSearchBarDelegate({required this.child});
+
+  @override
+  Widget build(
+    BuildContext context,
+    double shrinkOffset,
+    bool overlapsContent,
+  ) {
+    return Container(color: const Color(0xFF181A20), child: child);
+  }
+
+  @override
+  double get maxExtent => height;
+
+  @override
+  double get minExtent => height;
+
+  @override
+  bool shouldRebuild(covariant _SliverSearchBarDelegate oldDelegate) {
+    return false;
   }
 }
